refactoring panic_handler and add tests
