+++ reflectil/panic_handler_test.go
@@ -1,0 +1,23 @@
+package reflectil_test
+
+import (
+	"testing"
+
+	. "github.com/goldeneggg/structil/reflectil"
+)
+
+func TestRecoverToError(t *testing.T) {
+	t.Run("recover to error", func(t *testing.T) {
+		defer func() {
+			if r := recover(); r != nil {
+				err := RecoverToError(r)
+				if err == nil {
+					t.Errorf("RecoverToError() did not return error. r: %v", r)
+					return
+				}
+				t.Logf("%v", err)
+			}
+		}()
+		panic("panic for test")
+	})
+}
